add report probe stats sdk and command tool
