+++ config/manager-probe-agent/kustomization.yaml
@@ -1,5 +1,6 @@
 resources:
 - manager.yaml
+- rbac.yaml
 
 generatorOptions:
   disableNameSuffixHash: true
