+++ pkg/probe-status/status.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff"
+
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
