--- probers/dns-resolution-check/dns-resolution-check.yaml
@@ -1,32 +1,0 @@
-apiVersion: kubeprober.erda.cloud/v1
-kind: Probe
-metadata:
-  name: dns-resolution-check
-  namespace: kubeprober
-spec:
-  policy:
-    # unit: minute
-    runInterval: 2
-  probeList:
-    # probe item1:
-    - name: dns-resolution-check
-      spec:
-        containers:
-          - env:
-              - name: HOSTNAME
-                value: "kubernetes.default"
-              - name: NAMESPACE
-                value: "kube-system"
-              - name: DNS_NODE_SELECTOR
-                value: apps=k8s-dns
-              - name: NODE_NAME
-                valueFrom:
-                  fieldRef:
-                    fieldPath: spec.nodeName
-            name: dns-resolution-check
-            image: kubeprober/dns-resolution-check:v0.0.3
-            resources:
-              requests:
-                cpu: 10m
-                memory: 50Mi
-        restartPolicy: Never
