+++ probers/k8s/Dockerfile
@@ -1,0 +1,7 @@
+FROM kubeprober/probe-base:v0.0.1
+WORKDIR /
+
+COPY checker/ /checker/
+#USER 65532:65532
+
+CMD [ "sh", "-c", "/main.sh"]
